requirements.txt
openai==0.28.0
python-dotenv
tinydb
Flask
flask-cors

